botIO is now created via a factory. Also fixed a severe bug introduced by the previous commit
I'm not sure I'm that fond of botIO getting passed to like two thirds of all methods in the codebase, but it works and is technically no longer as "universal" (but still is) if that matters that much.

As for the bug: I tried removing several lines in UpdatesHandler that didn't seem to be doing anything productive, but that ended up breaking the bot (it could no longer recognize incoming messages). Still idk why. Further investigation may be needed.